backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
supabase==2.15.2
python-jose[cryptography]==3.4.0
pydantic-settings==2.9.1
httpx==0.28.1
resend==2.7.0
pytz==2025.2
python-dotenv==1.1.0
python-multipart==0.0.20
